+++ NEWS.md
@@ -1,3 +1,12 @@
+# nise 0.0.0.9001 (2022-08-30)
+
+Handle empty values
+
+- Modified `handle_nse_name()`: graceful handling of empty value (based on
+`character()`)
+
+-----
+
 # nise 0.0.0.9000 (2022-08-30)
 
 Initial commit